Added code to properly blow the parity fuse
By experimenting with the library i noticed that for some combination of FSG, SSG and OFFSET the output of the AD8556 will the fixed to GND no matter what the input were.
This problem was only present in programming mode, not in simulation mode.

After a little bit of research comparing the code and the datasheet, I noticed that the Parity fuse was never mentioned in the code and never blown accordingly to the specification. I added a variable (parity_counter) which keep track of the parity and allow the code to properly understand if the parity fuse has to be blown.
I also added a specific function to blow this fuse (blowParityFuse() ).